feat(voltron): rename voltron framework
