+++ framework/voltron/example/lib/view/base_voltron_page.dart
@@ -22,7 +22,7 @@ import 'dart:io';
 
 import 'package:device_info_plus/device_info_plus.dart';
 import 'package:flutter/material.dart';
-import 'package:voltron_render/voltron_render.dart';
+import 'package:voltron/voltron.dart';
 import 'package:voltron_renderer/voltron_renderer.dart';
 
 import '../my_api_provider.dart';
