+++ framework/voltron/ios/Classes/TencentVoltronRenderPlugin.m
@@ -21,13 +21,13 @@
  */
 
 #import "TencentVoltronRenderPlugin.h"
-#if __has_include(<voltron_render/voltron_render-Swift.h>)
-#import <voltron_render/voltron_render-Swift.h>
+#if __has_include(<voltron/voltron-Swift.h>)
+#import <voltron/voltron-Swift.h>
 #else
 // Support project import fallback if the generated compatibility header
 // is not copied when this plugin is created as a library.
 // https://forums.swift.org/t/swift-static-libraries-dont-copy-generated-objective-c-header/19816
-#import "voltron_render-Swift.h"
+#import "voltron-Swift.h"
 #endif
 
 @implementation TencentVoltronRenderPlugin
